change success message for marketplace apps
