+++ jumpscale/packages/marketplace/chats/taiga.py
@@ -143,21 +143,5 @@ class TaigaDeploy(MarketPlaceAppsChatflow):
             solutions.cancel_solution(self.user_info()["username"], self.workload_ids)
             raise StopChatFlow(f"Failed to create trc container on node {self.selected_node.node_id} {nginx_wid}")
 
-    @chatflow_step(title="Success", disable_previous=True, final_step=True)
-    def success(self):
-        self._wgconf_show_check()
-        message = f"""\
-# Taiga has been deployed successfully:
-\n<br />\n
-your reservation id is: `{self.workload_ids[1]}`
-\n<br />\n
-your container ip is: `{self.ip_address}`
-\n<br />\n
-open Taiga from browser at <a href="https://{self.domain}" target="_blank">https://{self.domain}</a>
-\n<br />\n
-- It may take few minutes to load.
-                """
-        self.md_show(message, md=True)
-
 
 chat = TaigaDeploy
